Made new component to build member preview rows based on given number
