+++ src/components/MemberPreviewBuilder/MemberPreviewBuilder.scss
@@ -1,0 +1,6 @@
+.member-preview-builder{
+  .member-preview-row{
+    display: flex;
+    margin-bottom: 50px;
+  }
+}
