+++ src/containers/HelpSomeonePage/HelpSomeonePage.scss
@@ -13,7 +13,7 @@ $general-text-color: #686868;
 		margin: 30px 0 50px 0;
 	}
 
-	p{
+	.intro-paragraph{
 		font-size: 20px;
 		margin-bottom: 40px;
 		color: $general-text-color;
@@ -36,11 +36,5 @@ $general-text-color: #686868;
 			font-weight: lighter;
 			font-size: 25px;
 		}
-	}
-
-	.member-preview-row{
-		display: flex;
-		justify-content: flex-start;
-		margin-bottom: 50px;
 	}
 }
